+++ components/aes70/src/aes70_internal.h
@@ -75,6 +75,7 @@ typedef enum {
     AES70_K_FILTER_PARAMETRIC,/* OcaFilterParametric (parametric EQ band) */
     AES70_K_PANBALANCE,       /* OcaPanBalance */
     AES70_K_SIGNAL_GEN,       /* OcaSignalGenerator */
+    AES70_K_GROUPER,          /* OcaGrouper (grouping agent) */
     AES70_K_DEVICE_MANAGER,
     AES70_K_SUBSCRIPTION_MANAGER,
     AES70_K_COUNT
@@ -264,6 +265,16 @@ aes70_status_t aes70_root_dispatch(struct aes70_object *obj, uint16_t idx,
                                    ocp1_rd_t *in, ocp1_wr_t *out, uint8_t *pc);
 aes70_status_t aes70_worker_dispatch(struct aes70_object *obj, uint16_t idx,
                                      ocp1_rd_t *in, ocp1_wr_t *out, uint8_t *pc);
+aes70_status_t aes70_agent_dispatch(struct aes70_object *obj, uint16_t idx,
+                                    ocp1_rd_t *in, ocp1_wr_t *out, uint8_t *pc);
+
+/* ---- OcaGrouper (aes70_grouper.c) --------------------------------------- */
+aes70_status_t aes70_grouper_dispatch(struct aes70_object *obj, uint16_t idx,
+                                      ocp1_rd_t *in, ocp1_wr_t *out, uint8_t *pc);
+void aes70_grouper_init(struct aes70_object *obj, aes70_kind_t citizen_kind);
+void aes70_grouper_on_commit(aes70_device_t *dev, struct aes70_object *written);
+/* The (level,index) of an object's primary value property; false if none. */
+bool aes70_object_primary_property(struct aes70_object *obj, uint16_t *level, uint16_t *index);
 
 /* ---- Identification actuator (aes70_classes.c) -------------------------- */
 aes70_status_t aes70_identify_dispatch(struct aes70_object *obj, uint16_t idx,